Spark3.2.2WithJarsUpdate

+++ sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/TPCDSQueryBenchmark.scala
@@ -23,14 +23,12 @@ import org.apache.spark.SparkConf
 import org.apache.spark.benchmark.Benchmark
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.TPCDSSchema
 import org.apache.spark.sql.catalyst.catalog.HiveTableRelation
 import org.apache.spark.sql.catalyst.plans.logical.SubqueryAlias
 import org.apache.spark.sql.catalyst.util._
 import org.apache.spark.sql.catalyst.util.DateTimeConstants.NANOS_PER_SECOND
 import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.types.StructType
 
 /**
  * Benchmark to measure TPCDS query performance.
@@ -39,9 +37,9 @@ import org.apache.spark.sql.types.StructType
  *   1. without sbt:
  *        bin/spark-submit --jars <spark core test jar>,<spark catalyst test jar>
  *          --class <this class> <spark sql test jar> --data-location <location>
- *   2. build/sbt "sql/Test/runMain <this class> --data-location <TPCDS data location>"
+ *   2. build/sbt "sql/test:runMain <this class> --data-location <TPCDS data location>"
  *   3. generate result: SPARK_GENERATE_BENCHMARK_FILES=1 build/sbt
- *        "sql/Test/runMain <this class> --data-location <location>"
+ *        "sql/test:runMain <this class> --data-location <location>"
  *      Results will be written to "benchmarks/TPCDSQueryBenchmark-results.txt".
  * }}}
  */
@@ -67,19 +65,23 @@ object TPCDSQueryBenchmark extends SqlBasedBenchmark with Logging {
     "web_returns", "web_site", "reason", "call_center", "warehouse", "ship_mode", "income_band",
     "time_dim", "web_page")
 
-  def setupTables(dataLocation: String, tableColumns: Map[String, StructType]): Map[String, Long] =
+  def setupTables(dataLocation: String, createTempView: Boolean): Map[String, Long] = {
     tables.map { tableName =>
-      spark.sql(s"DROP TABLE IF EXISTS $tableName")
-      val options = Map("path" -> s"$dataLocation/$tableName")
-      spark.catalog.createTable(tableName, "parquet", tableColumns(tableName), options)
-      // Recover partitions but don't fail if a table is not partitioned.
-      Try {
-        spark.sql(s"ALTER TABLE $tableName RECOVER PARTITIONS")
-      }.getOrElse {
-        logInfo(s"Recovering partitions of table $tableName failed")
+      if (createTempView) {
+        spark.read.parquet(s"$dataLocation/$tableName").createOrReplaceTempView(tableName)
+      } else {
+        spark.sql(s"DROP TABLE IF EXISTS $tableName")
+        spark.catalog.createTable(tableName, s"$dataLocation/$tableName", "parquet")
+        // Recover partitions but don't fail if a table is not partitioned.
+        Try {
+          spark.sql(s"ALTER TABLE $tableName RECOVER PARTITIONS")
+        }.getOrElse {
+          logInfo(s"Recovering partitions of table $tableName failed")
+        }
       }
       tableName -> spark.table(tableName).count()
     }.toMap
+  }
 
   def runTpcdsQueries(
       queryLocation: String,
@@ -161,7 +163,7 @@ object TPCDSQueryBenchmark extends SqlBasedBenchmark with Logging {
     }
 
     val tableSizes = setupTables(benchmarkArgs.dataLocation,
-      TPCDSSchemaHelper.getTableColumns)
+      createTempView = !benchmarkArgs.cboEnabled)
     if (benchmarkArgs.cboEnabled) {
       spark.sql(s"SET ${SQLConf.CBO_ENABLED.key}=true")
       spark.sql(s"SET ${SQLConf.PLAN_STATS_ENABLED.key}=true")
@@ -183,9 +185,4 @@ object TPCDSQueryBenchmark extends SqlBasedBenchmark with Logging {
     runTpcdsQueries(queryLocation = "tpcds-v2.7.0", queries = queriesV2_7ToRun, tableSizes,
       nameSuffix = nameSuffixForQueriesV2_7)
   }
-}
-
-object TPCDSSchemaHelper extends TPCDSSchema {
-  def getTableColumns: Map[String, StructType] =
-    tableColumns.map(kv => kv._1 -> StructType.fromDDL(kv._2))
 }
